Move `RejectedPromisesTracker` to its own module
